\name{BA23001021}
\alias{BA23001021}
\docType{package}
\title{
  Multiple time change-points detection in Cox model by dynamic programming
}
\description{
  Use dynamic programming to detect time change-points in Cox model.
}
\author{
Yulin Zhou
Maintainer: Yulin Zhou <zhouyulin@mail.ustc.edu.cn>
}
\keyword{ package }
