% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plik.r
\name{dp.mcp}
\alias{dp.mcp}
\title{Detect multiple change-points by dynamic programming.}
\usage{
dp.mcp(t, delta, Z, candid, ac)
}
\arguments{
\item{t}{Survival time}

\item{delta}{Censoring indicator (1 failure, 0 right censoring)}

\item{Z}{Covariates}

\item{candid}{Candid time points}

\item{ac}{1 - significant level}
}
\value{
A list containing all information
}
\description{
Use dynamic programming to detect multiple change-points
}
