% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plik.r
\name{plik.1d}
\alias{plik.1d}
\title{Derivative of log partial-likelihood.}
\usage{
plik.1d(t, delta, Z, beta)
}
\arguments{
\item{t}{Survival time}

\item{delta}{Censoring indicator (1 failure, 0 right censoring)}

\item{Z}{Covariates}

\item{beta}{Coefficients}
}
\value{
Derivative of log partial-likelihood (numeric vector)
}
\description{
Compute derivative of log partial-likelihood under Cox model, assume right censoring and no tie
}
\examples{
\dontrun{
t <- rexp(1e3); delta <- rep(1, 1e3)
Z <- rnorm(1e3); beta <- 1
plik.1d(t, delta, Z, beta)
}
}
