% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plik.r
\name{plik.est}
\alias{plik.est}
\title{Partial-likelihood estimation.}
\usage{
plik.est(t, delta, Z, init = rep(0, NCOL(Z)), control = 1e-05)
}
\arguments{
\item{t}{Survival time}

\item{delta}{Censoring indicator (1 failure, 0 right censoring)}

\item{Z}{Covariates}

\item{init}{Initial value}

\item{control}{Convergence threshold}
}
\value{
A list containing the estimation and the variance-covariance matrix
}
\description{
Use Newton-Raphson method to compute maximum partial-likelihood estimation, assume right censoring and no tie
}
\examples{
\dontrun{
t <- rexp(1e3); delta <- rep(1, 1e3)
Z <- rnorm(1e3)
plik.est(t, delta, Z)
}
}
