% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plik.r
\name{plik}
\alias{plik}
\title{Log partial-likelihood.}
\usage{
plik(t, delta, Z, beta)
}
\arguments{
\item{t}{Survival time}

\item{delta}{Censoring indicator (1 failure, 0 right censoring)}

\item{Z}{Covariates}

\item{beta}{Coefficients}
}
\value{
Log partial-likelihood (numeric)
}
\description{
Compute log partial-likelihood under Cox model, assume right censoring and no tie
}
\examples{
\dontrun{
t <- rexp(1e3); delta <- rep(1, 1e3)
Z <- rnorm(1e3); beta <- 1
plik(t, delta, Z, beta)
}
}
