% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plik.r
\name{simdata}
\alias{simdata}
\title{A dataset used for illustration.}
\description{
This dataset is simulated, with 2 time change-points at 3, 6
}
\examples{
\dontrun{
data(simdata)
t <- simdata$t; delta <- simdata$delta; Z <- simdata$z
plik.est(t, delta, Z)
}
}
